Add prettier

+++ contracts/Migrations.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 contract Migrations {
   address public owner;
-  uint public lastCompletedMigration;
+  uint256 public lastCompletedMigration;
 
   constructor() {
     owner = msg.sender;
@@ -14,7 +14,7 @@ contract Migrations {
     _;
   }
 
-  function setCompleted(uint completed) external restricted {
+  function setCompleted(uint256 completed) external restricted {
     lastCompletedMigration = completed;
   }
